Update SMAv4 to latest version
